requirements-streamlit.txt
# Streamlit explorer prototype (refs https://github.com/jimchurches/myebirdstuff/issues/70)
# pip install -r requirements-streamlit.txt
# streamlit run streamlit_app/app.py
#
# Pandas: Streamlit currently declares pandas<3 (e.g. 1.55.x). The Jupyter explorer uses
# pandas>=3 in requirements-explorer.txt — use a separate virtualenv for Streamlit if you
# need both, or pip will downgrade pandas in a shared env.

# Python 3.10+ (streamlit-searchbox; refs #70)
# 1.40+: ``st.dataframe(..., width="stretch")`` (``use_container_width`` deprecated / removed later).
streamlit>=1.40
pandas>=2.0,<3
folium>=0.15
streamlit-folium>=0.20.0
# Maintenance tab: close-location pairs (BallTree / haversine); refs #69, #79
scikit-learn>=1.3,<2
# Country tab + rankings: ISO codes → display names (``region_display.country_for_display``)
pycountry>=23.12.11
# Species autocomplete (Python >= 3.10); refs #70
whoosh
streamlit-searchbox>=0.1.24
pydantic>=2,<3
PyYAML>=6,<7
